separate TInherit from Any.h
