+++ Source/Mcro/Public/Mcro/CommonCore.h
@@ -32,6 +32,7 @@
 #include "Mcro/Finally.h"
 #include "Mcro/FmtMacros.h"
 #include "Mcro/FunctionTraits.h"
+#include "Mcro/Inheritance.h"
 #include "Mcro/InitializeOnCopy.h"
 #include "Mcro/Once.h"
 #include "Mcro/SharedObjects.h"
@@ -59,6 +60,7 @@ namespace Mcro::CommonCore
 	using namespace Mcro::Enums;
 	using namespace Mcro::Finally;
 	using namespace Mcro::FunctionTraits;
+	using namespace Mcro::Inheritance;
 	using namespace Mcro::InitializeOnCopy;
 	using namespace Mcro::Once;
 	using namespace Mcro::SharedObjects;
@@ -67,6 +69,7 @@ namespace Mcro::CommonCore
 	using namespace Mcro::Types;
 	using namespace Mcro::Range;
 	using namespace Mcro::ValueThunk;
+	using namespace Mcro::Zero;
 }
 
 /** @brief Use this namespace for all the common features MCRO has to offer */
